alx_travel_app/requirement.txt
Django
djangorestframework
django-cors-headers
django-environ
psycopg2-binary
dj-database-url
drf-spectacular
django-filter
redis